locktorture: More elaborate testing of jbd2 locking and refcounting
This modifies writers to use the b_jcount as refcounting. Every writer
will increase the b_jcount under the lock, unlock and after some short
wait acquire the lock and decrease the b_jcount. All that while making
sure that the b_jcount does not drop to zero or below.

Set the b_jcount to 1 on initialization and test that on cleanup.

Additionally introduce readers that do not use the jbd2 bit_spin_lock,
but rather a regular spin_lock and use set_bit() on bh->b_state to set
and clear random flags (bits 0-25 which is actually used in real code).

Because the cleanup test is very important as well I run the torture
with the following script

trap cleanup 0 1 2 3 15

cleanup() {
        modprobe -r locktorture
        exit 0
}

loop=0
while true; do
        loop=$((loop+1))
        modprobe locktorture torture_type=jbd2_bit_spin_lock nwriters_stress=100 stutter=0 verbose=0 nreaders_stress=20
        echo -n "Loop $loop ... "
        sleep 20m
        echo "DONE"
        modprobe -r locktorture
        sleep 1
done

Signed-off-by: Lukas Czerner <lczerner@redhat.com>